// src/manager/api/sgp/types.ts
export interface AvailableServersMap {
  servers: {
    [region: string]: {
      /**
       * 服务器名称
       */
      name: string

      /**
       * 用于战绩查询的服务器地址
       */
      matchHistory: string

      /**
       * 其他通用查询服务器地址
       */
      common: string
    }
  }

  /**
   * 腾讯服务器中，以下服务器可以使用同一个 jwt token 实现战绩查询
   */
  tencentServerMatchHistoryInteroperability: string[]

  /**
   * 腾讯服务器中，以下服务器可以使用同一个 jwt token 实现观战查询
   */
  tencentServerSpectatorInteroperability: string[]

  /**
   * 腾讯服务器中，以下服务器可以使用同一个 jwt token 实现召唤师查询
   */
  tencentServerSummonerInteroperability: string[]
}

export interface SgpMatchHistoryLol {
  games: SgpGameSummaryLol[]
}

export interface SgpGameSummaryLol {
  metadata: SgpGameMetadataLol
  json: SgpGameSummaryJsonLol
}

export interface SgpGameSummaryJsonLol {
  endOfGameResult: string
  gameCreation: number
  gameDuration: number
  gameEndTimestamp: number
  gameId: number
  gameMode: string
  gameName: string
  gameStartTimestamp: number
  gameType: string
  gameVersion: string
  mapId: number
  participants: SgpParticipantLol[]
  platformId: string
  queueId: number
  seasonId: number
  teams: Team[]
  tournamentCode: string
}

export interface Team {
  bans: Ban[]
  objectives: Objectives
  teamId: number
  win: boolean
}

interface Objectives {
  baron: Baron
  champion: Baron
  dragon: Baron
  horde: Baron
  inhibitor: Baron
  riftHerald: Baron
  tower: Baron
}

interface Baron {
  first: boolean
  kills: number
}

interface Ban {
  championId: number
  pickTurn: number
}

export interface SgpParticipantLol {
  allInPings: number
  assistMePings: number
  assists: number
  baronKills: number
  basicPings: number
  bountyLevel: number
  challenges: Challenges
  champExperience: number
  champLevel: number
  championId: number
  championName: string
  championTransform: number
  commandPings: number
  consumablesPurchased: number
  damageDealtToBuildings: number
  damageDealtToObjectives: number
  damageDealtToTurrets: number
  damageSelfMitigated: number
  dangerPings: number
  deaths: number
  detectorWardsPlaced: number
  doubleKills: number
  dragonKills: number
  eligibleForProgression: boolean
  enemyMissingPings: number
  enemyVisionPings: number
  firstBloodAssist: boolean
  firstBloodKill: boolean
  firstTowerAssist: boolean
  firstTowerKill: boolean
  gameEndedInEarlySurrender: boolean
  gameEndedInSurrender: boolean
  getBackPings: number
  goldEarned: number
  goldSpent: number
  holdPings: number
  individualPosition: string
  inhibitorKills: number
  inhibitorTakedowns: number
  inhibitorsLost: number
  item0: number
  item1: number
  item2: number
  item3: number
  item4: number
  item5: number
  item6: number
  itemsPurchased: number
  killingSprees: number
  kills: number
  lane: string
  largestCriticalStrike: number
  largestKillingSpree: number
  largestMultiKill: number
  longestTimeSpentLiving: number
  magicDamageDealt: number
  magicDamageDealtToChampions: number
  magicDamageTaken: number
  missions: SgpMissions
  needVisionPings: number
  neutralMinionsKilled: number
  nexusKills: number
  nexusLost: number
  nexusTakedowns: number
  objectivesStolen: number
  objectivesStolenAssists: number
  onMyWayPings: number
  participantId: number
  pentaKills: number
  perks: Perks
  physicalDamageDealt: number
  physicalDamageDealtToChampions: number
  physicalDamageTaken: number
  placement: number
  playerAugment1: number
  playerAugment2: number
  playerAugment3: number
  playerAugment4: number
  playerAugment5: number
  playerAugment6: number
  playerSubteamId: number
  profileIcon: number
  pushPings: number
  puuid: string
  quadraKills: number
  riotIdGameName: string
  riotIdTagline: string
  role: string
  sightWardsBoughtInGame: number
  spell1Casts: number
  spell1Id: number
  spell2Casts: number
  spell2Id: number
  spell3Casts: number
  spell4Casts: number
  subteamPlacement: number
  summoner1Casts: number
  summoner2Casts: number
  summonerId: number
  summonerLevel: number
  summonerName: string
  teamEarlySurrendered: boolean
  teamId: number
  teamPosition: string
  timeCCingOthers: number
  timePlayed: number
  totalAllyJungleMinionsKilled: number
  totalDamageDealt: number
  totalDamageDealtToChampions: number
  totalDamageShieldedOnTeammates: number
  totalDamageTaken: number
  totalEnemyJungleMinionsKilled: number
  totalHeal: number
  totalHealsOnTeammates: number
  totalMinionsKilled: number
  totalTimeCCDealt: number
  totalTimeSpentDead: number
  totalUnitsHealed: number
  tripleKills: number
  trueDamageDealt: number
  trueDamageDealtToChampions: number
  trueDamageTaken: number
  turretKills: number
  turretTakedowns: number
  turretsLost: number
  unrealKills: number
  visionClearedPings: number
  visionScore: number
  visionWardsBoughtInGame: number
  wardsKilled: number
  wardsPlaced: number
  win: boolean
}
// ...
export interface SgpGameMetadataLol {
  product: string
  tags: string[]
  participants: string[]
  timestamp: string
  data_version: string
  info_type: string
  match_id: string
  private: boolean
}
// ...
export interface SgpGameDetailsLol {
  metadata: SgpGameMetadataLol
  json: SgpGameJsonDetailsLol
}

interface SgpGameJsonDetailsLol {
  endOfGameResult: string
  frameInterval: number
  frames: Frame[]
  gameId: number
  participants: Participant[]
}

interface Participant {
  participantId: number
  puuid: string
}

interface Frame {
  events: Event[]
  participantFrames: ParticipantFrames
  timestamp: number
}
// ...
export interface SgpRankedStats {
  queues: Queue[]
  highestPreviousSeasonEndTier: string
  highestPreviousSeasonEndRank: string
  highestPreviousSeasonAchievedTier: string
  highestPreviousSeasonAchievedRank: string
  earnedRegaliaRewardIds: any[]
  currentSeasonSplitPoints: number
  previousSeasonSplitPoints: number
  seasons: Seasons
}

interface Seasons {
  RANKED_TFT: RANKEDTFT
  RANKED_TFT_TURBO: RANKEDTFT
  CHERRY: RANKEDTFT
  RANKED_FLEX_SR: RANKEDTFT
  RANKED_TFT_DOUBLE_UP: RANKEDTFT
  RANKED_SOLO_5x5: RANKEDTFT
}

interface RANKEDTFT {
  currentSeasonId: number
  currentSeasonEnd: number
  nextSeasonStart: number
}

interface Queue {
  queueType: string
  provisionalGameThreshold: number
  tier?: string
  rank?: string
  leaguePoints: number
  cumulativeLp: number
  wins: number
  losses: number
  provisionalGamesRemaining: number
  highestTier?: string
  highestRank?: string
  previousSeasonEndTier?: string
  previousSeasonEndRank?: string
  previousSeasonHighestTier?: string
  previousSeasonHighestRank?: string
  previousSeasonAchievedTier?: string
  previousSeasonAchievedRank?: string
  ratedRating: number
  premadeMmrRestricted: boolean
  ratedTier?: string
}

export interface SgpSummoner {
  id: number
  puuid: string
  accountId: number
  name: string
  internalName: string
  profileIconId: number
  level: number
  expPoints: number
  levelAndXpVersion: number
  revisionId: number
  revisionDate: number
  lastGameDate: number
  nameChangeFlag: boolean
  unnamed: boolean
  privacy: string
  expToNextLevel: number
}


export interface SpectatorData {
  reconnectDelay: number
  gameName: string
  game: SpectatorGameflowSession
  playerCredentials: SpectatorPlayerCredentials
}

interface SpectatorPlayerCredentials {
  gameId: number
  queueId: number
  playerId: number
  puuid: string
  serverPort: number
  championId: number
  lastSelectedSkinIndex: number
  summonerId: number
  observer: boolean
  gameVersion: string
  gameMode: string
  observerEncryptionKey: string
  observerServerIp: string
  observerServerPort: number
  queueType: string
  gameCreateDate: number
}

interface SpectatorGameflowSession {
  id: number
  gameState: string
  queueTypeName: string
  name: string
  pickTurn: number
  mapId: number
  gameMode: string
  maxNumPlayers: number
  gameType: string
  gameQueueConfigId: number
  spectatorDelay: number
  gameVersion: string
  teamOne: SpectatorGameflowSessionTeam[]
  teamTwo: SpectatorGameflowSessionTeam[]
  playerChampionSelections: SpectatorGameflowSessionPlayerChampionSelection[]
  bannedChampions: any[]
  observers: any[]
}

interface SpectatorGameflowSessionPlayerChampionSelection {
  summonerInternalName: string
  championId: number
  selectedSkinIndex: number
  spell1Id: number
  spell2Id: number
}

interface SpectatorGameflowSessionTeam {
  puuid: string
  summonerId: number
  lastSelectedSkinIndex: number
  teamOwner: boolean
  profileIconId: number
  teamParticipantId: number
  championId: number
  selectedRole: string
  selectedPosition: string
  summonerName: string
  summonerInternalName: string
}